Update travis build
For some reason, Travis CI moved everything from travis-ci.org to travis-ci.com and put everyone else through a lot of extra work to follow the move. So, I just went through the work for this repo. Now travis-ci.com should now build for pushes and pull requests, and this PR will both test that out and update the readme with the new build status badge.